deafult implementation of compile_hint
