+++ src/hint_processor/builtin_hint_processor/builtin_hint_processor_definition.rs
@@ -1,4 +1,3 @@
-use crate::any_box;
 use crate::hint_processor::builtin_hint_processor::blake2s_utils::{
     blake2s_add_uint256, blake2s_add_uint256_bigend, compute_blake2s, finalize_blake2s,
 };
@@ -438,41 +437,6 @@ impl HintProcessor for BuiltinHintProcessor {
             code => Err(VirtualMachineError::UnknownHint(code.to_string())),
         }
     }
-
-    fn compile_hint(
-        &self,
-        code: &str,
-        ap_tracking: &ApTracking,
-        reference_ids: &HashMap<String, usize>,
-        references: &HashMap<usize, HintReference>,
-    ) -> Result<Box<dyn Any>, VirtualMachineError> {
-        Ok(any_box!(HintProcessorData {
-            code: code.to_string(),
-            ap_tracking: ap_tracking.clone(),
-            ids_data: get_ids_data(reference_ids, references)?,
-        }))
-    }
-}
-
-fn get_ids_data(
-    reference_ids: &HashMap<String, usize>,
-    references: &HashMap<usize, HintReference>,
-) -> Result<HashMap<String, HintReference>, VirtualMachineError> {
-    let mut ids_data = HashMap::<String, HintReference>::new();
-    for (path, ref_id) in reference_ids {
-        let name = path
-            .rsplit('.')
-            .next()
-            .ok_or(VirtualMachineError::FailedToGetIds)?;
-        ids_data.insert(
-            name.to_string(),
-            references
-                .get(ref_id)
-                .ok_or(VirtualMachineError::FailedToGetIds)?
-                .clone(),
-        );
-    }
-    Ok(ids_data)
 }
 
 #[cfg(test)]
